Revert "feat: migrate dust apps v1 endpoints to vaults (#7260)"
This reverts commit e21329365c188a7093cb8b92bdf25b0b4e510bae.
